requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.1.0  # Excel file support

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine learning
scikit-learn>=1.2.0

# Jupyter notebooks
jupyter>=1.0.0
notebook>=6.5.0

# Testing
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0